@@ -22,40 +22,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "^0.2.0",
-    "@angular/animations": "5.1.0",
-    "@angular/cdk": "5.0.3",
-    "@angular/cli": "github:nrwl/fix-cli-build#release16",
-    "@angular/common": "5.1.0",
-    "@angular/compiler": "5.1.0",
-    "@angular/compiler-cli": "5.1.0",
-    "@angular/core": "5.1.0",
-    "@angular/flex-layout": "^5.0.0-beta.13",
-    "@angular/forms": "5.1.0",
-    "@angular/http": "5.1.0",
-    "@angular/language-service": "5.1.0",
-    "@angular/material": "5.0.3",
-    "@angular/platform-browser": "5.1.0",
-    "@angular/platform-browser-dynamic": "5.1.0",
-    "@angular/router": "5.1.0",
-    "@ngrx/effects": "^4.1.1",
-    "@ngrx/entity": "^4.1.1",
-    "@ngrx/router-store": "^4.1.1",
-    "@ngrx/store": "4.1.1",
-    "@ngrx/store-devtools": "4.1.1",
-    "@nrwl/nx": "0.5.2",
+    "@angular/animations": "^6.0.0",
+    "@angular-devkit/core": "0.3.2",
+    "@angular/cli": "^1.7.4",
+    "@angular/cdk": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/language-service": "^6.0.0",
+    "@angular/core": "^6.0.0", 
+    "@angular/flex-layout": "^6.0.0-beta.15",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/material": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
+    "@ngrx/effects": "4.1.1",
+    "@ngrx/entity": "^6.0.0-beta.1",
+    "@ngrx/router-store": "^6.0.0-beta.1",
+    "@ngrx/store": "^6.0.0-beta.1",
+    "@ngrx/store-devtools": "^6.0.0-beta.1",
+    "@nrwl/nx": "^6.0.0",
     "@nrwl/schematics": "0.5.2",
     "@swimlane/ngx-charts": "^7.0.1",
     "@tinymce/tinymce-angular": "1.0.2",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
+    "@types/jwt-decode": "^2.2.1",
+    "@types/lodash": "^4.14.106",
+    "@types/node": "^9.6.6",
+    "angular-tslint-rules": "^1.2.1",
+    "codelyzer": "^4.2.1",
     "angular-gridster2": "4.7.1",
     "angular-handsontable": "^1.0.0-beta3",
     "angular2-froala-wysiwyg": "2.6.0",
     "angular2-grid": "2.1.0",
     "babyparse": "0.4.6",
-    "codelyzer": "~3.1.1",
     "copy-webpack-plugin": "^4.3.1",
     "core-js": "^2.4.1",
     "d3": "4.9.1",
@@ -66,6 +69,7 @@
     "dom-to-image": "^2.6.0",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
+    "jsrsasign": "^8.0.10",
     "karma": "~1.7.0",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
@@ -76,26 +80,35 @@
     "ng2-dnd": "5.0.2",
     "ng2-dragula": "1.5.0",
     "ng4-codemirror": "0.0.1",
-    "ngrx-store-freeze": "^0.2.0",
     "ngx-bootstrap-feedback": "^1.1.5",
     "ngx-color-picker": "^5.3.2",
+    "ngrx-store-freeze": "^0.2.2",
     "ngx-uploader": "^4.2.2",
+    "rxjs-compat": "^6.0.0",
+    "rxjs-tslint": "^0.1.4",
+    "rxjs-tslint-rules": "^4.0.2",
+    "object-path": "^0.11.4",
+    "ts-node": "^3.2.2",
+    "tslint": "^5.9.1",
+    "tslint-angular": "^1.1.1",
+    "tslint-config-airbnb": "^5.8.0",
+    "typescript": "2.7.2",
     "prettier": "1.8.2",
     "protractor": "~5.1.2",
-    "rxjs": "^5.5.2",
+    "rxjs": "^6.0.0",
     "screenfull": "^3.3.2",
     "tinymce": "4.7.6",
-    "ts-node": "~3.2.0",
     "ts-xlsx": "0.0.11",
-    "tslint": "~5.3.2",
-    "typescript": ">=2.4.2 <2.5.0",
-    "zone.js": "^0.8.14"
+    "zone.js": "^0.8.14",
+    "angular-auth-oidc-client": "^6.0.0",
+    "angular-google-signin": "^0.1.5",
+    "hammerjs": "^2.0.8"
   },
   "devDependencies": {
     "@angular/cli": "github:nrwl/fix-cli-build#release16",
     "@angular/compiler-cli": "5.1.3",
     "@angular/language-service": "5.1.3",
-    "@ngrx/store-devtools": "^4.1.1",
+    "@ngrx/store-devtools": "^6.0.0-beta.1",
     "@nrwl/schematics": "0.5.2",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
