@@ -1,5 +1,21 @@
-export { LogoutDirective } from './src/directives/logout/logout.directive';
+export { AuthenticationRoutingModule } from './src/authentication-routing.module';
+export { AuthenticationStateModule } from './src/authentication-state.module';
 export { AuthenticationGuardService } from './src/services/authentication.guard.service';
-export { AuthenticationComponent } from './src/components/authentication/authentication.component';
-export { AuthenticationModule } from './src/authentication.module';
-export { ResetPasswordComponent } from './src/components/reset-password/reset-password.component';
+export { RoleGuardService } from './src/services/role.guard.service';
+import * as fromAuthentication from './src/+state/actions/authentication-state.actions';
+export { fromAuthentication };
+export {
+  getUser,
+  getLoggedIn,
+  getIsUserLoading,
+  getLoginPagePending,
+  getLoginPageError,
+  selectIsLoggedIn,
+  selectUser,
+  selectLoginPageError,
+  selectLoginPagePending
+} from './src/+state/selectors/authentication-state.selectors';
+export { AuthenticationState } from './src/+state/states/authentication-state.state';
+export { authenticationReducers } from './src/+state/reducers/authentication-state.reducers';
+export { User } from './src/models/user.model';
+export { AuthenticationViewModule } from './src/authentication-view.module';
