@@ -1,58 +1,42 @@
 import { Store } from '@ngrx/store';
-import {
-  HttpResponse,
-  HttpRequest,
-  HttpHandler,
-  HttpEvent,
-  HttpInterceptor,
-  HttpErrorResponse
-} from '@angular/common/http';
+import { HttpErrorResponse, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Router } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 import { empty } from 'rxjs/observable/empty';
-import { switchMap } from 'rxjs/operators/switchMap';
-import { first } from 'rxjs/operators/first';
 import { _throw } from 'rxjs/observable/throw';
 import { catchError } from 'rxjs/operators/catchError';
-import { fromAuthentication, selectTokenExpiresIn, AuthenticationState } from '@labdat/authentication-state';
-import { isEmpty } from 'lodash';
+import { AuthenticationState } from '../+state/states/authentication-state.state';
+import * as fromAuthentication from '../+state/actions/authentication-state.actions';
 
 @Injectable()
 export class AuthenticationInterceptorService implements HttpInterceptor {
-  private urlFilters = ['/api'];
-  private tokenExpiresIn$;
 
-  constructor(private store: Store<AuthenticationState>) {
-    this.tokenExpiresIn$ = this.store.select(selectTokenExpiresIn);
-  }
+  constructor(private store: Store<AuthenticationState>) { }
 
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<any> {
-    if (!/\/api\//.exec(request.url)) {
-      return next.handle(request);
+    if (request.url.match(/(?!\api\/)/)) {
+      const nextRequest = request.clone({ withCredentials: true });
+
+      return next.handle(nextRequest);
     }
 
-    return this.tokenExpiresIn$.pipe(
-      first(),
-      switchMap(tokenExpiresIn => {
-        if (tokenExpiresIn && tokenExpiresIn < Date.now()) {
-          this.store.dispatch(new fromAuthentication.Logout('Token Expired'));
+    const tokenExpiresIn = Number(localStorage.getItem('tokenExpiresIn'));
+    if (tokenExpiresIn && tokenExpiresIn < Date.now()) {
+      this.store.dispatch(new fromAuthentication.RemoteLogout('Token expired'));
+
+      return empty();
+    }
+
+    return next.handle(request)
+    .pipe(
+      catchError((error: any) => {
+        if (error instanceof HttpErrorResponse && error.status === 403 && !error.url.includes('/auth')) {
+          this.store.dispatch(new fromAuthentication.RemoteLogout('Unauthorized Operation'));
+
           return empty();
         }
 
-        const req = request.clone({
-          withCredentials: true
-        });
-
-        return next.handle(req).pipe(
-          catchError(error => {
-            if (error instanceof HttpErrorResponse && error.status === 403 && !error.url.includes('/auth')) {
-              this.store.dispatch(new fromAuthentication.Logout('Unauthorized Operation'));
-              return empty();
-            }
-            return _throw(error);
-          })
-        );
+        return _throw(error);
       })
     );
   }
