@@ -1,4 +1,5 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { MatCardModule } from '@angular/material/card';
 
 import { HomeComponent } from './home.component';
 
@@ -9,8 +10,12 @@ describe('HomeComponent', () => {
   beforeEach(
     async(() => {
       TestBed.configureTestingModule({
+        imports: [
+          MatCardModule
+        ],
         declarations: [HomeComponent]
-      }).compileComponents();
+      })
+      .compileComponents();
     })
   );
 
@@ -21,6 +26,7 @@ describe('HomeComponent', () => {
   });
 
   it('should create', () => {
-    expect(component).toBeTruthy();
+    expect(component)
+    .toBeTruthy();
   });
 });
