@@ -1,6 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { NotFoundComponent } from './not-found.component';
+import { MatCardModule } from '@angular/material/card';
 
 describe('NotFoundComponent', () => {
   let component: NotFoundComponent;
@@ -9,8 +10,12 @@ describe('NotFoundComponent', () => {
   beforeEach(
     async(() => {
       TestBed.configureTestingModule({
+        imports: [
+          MatCardModule
+        ],
         declarations: [NotFoundComponent]
-      }).compileComponents();
+      })
+      .compileComponents();
     })
   );
 
@@ -21,6 +26,7 @@ describe('NotFoundComponent', () => {
   });
 
   it('should create', () => {
-    expect(component).toBeTruthy();
+    expect(component)
+    .toBeTruthy();
   });
 });
