@@ -1,16 +1,16 @@
 import { MenuItem } from '@labdat/data-models';
 
 interface PresentationsConfiguration {
-  core: MenuItem[];
+  core: { sidenav: Array<MenuItem> };
 }
 
 export const presentationsConfiguration: PresentationsConfiguration = {
-  core: [
-    {
+  core: {
+    sidenav: [{
       order: 4,
       link: '/presentations',
       name: 'Presentations',
       icon: 'hardware:ic_desktop_mac_24px'
-    }
-  ]
-};
+    }]
+  }
+}
