@@ -20,14 +20,14 @@ import { debounceTime } from 'rxjs/operators/debounceTime';
 import { switchMap } from 'rxjs/operators/switchMap';
 import { tap } from 'rxjs/operators/tap';
 import { Store } from '@ngrx/store';
-import { selectIsLoggedIn, selectUser } from '@labdat/authentication-state';
+import { selectIsLoggedIn, selectUser } from '@labdat/authentication';
 import { selectSlideIds, selectAllSlides, SlidesState } from '@labdat/slides-state';
 import { selectCurrentPresentationId } from '@labdat/slides-state';
 import { fromSlides } from '@labdat/slides-state';
 import { withLatestFrom } from 'rxjs/operators/withLatestFrom';
 import { Subject } from 'rxjs/Subject';
 import { Subscription } from 'rxjs/Subscription';
-import { fromRouter } from '@labdat/router-state';
+import { fromRouter } from '@labdat/common/router-state';
 import { MatDialog } from '@angular/material/dialog';
 import { SlideDialogComponent } from '../../components/slide-dialog/slide-dialog.component';
 import { zip } from 'rxjs/observable/zip';
